Discount System added.
Now we have discount tables which has Many to Many relationships with the users table.
Also now product_lines table has a foreign key to discounts table as well.

The new method 'getFinalPrice()' has been arranged for ProductLine model to return
the price considering the discount, if there is one.