The Ajax for handling Modal changed completely: Now instead of adding HTML records to the blade file, sends a new get request to the same page.
New Image Upload System configured;
Now There is a ImageUploadService Class that handles image uploads.
With this service , now when a Admin User creates a new Product Line, can also
attach an Image to that instance.
The Image attachment proccess of the Post creation changed as well to the new method.

 On branch Admin-Panel

 Changes to be committed:
	modified:   app/Http/Controllers/Admin/BlogController.php
	modified:   app/Http/Controllers/Admin/ProductController.php
	modified:   app/Models/Images/PostImage.php
	new file:   app/Support/ImageUpload/ImageUploadService.php
	new file:   public/js/select.js
	modified:   resources/views/admin/product/productList.blade.php